feat: add support for continuing work across chats
Let's add a way to continue working on a chat in a new chat. To do this, we need an optional argument to InitProject that instructs it to reuse the chat id of the HEAD commit. We also need to adjust the prompt in docblock for main.py to set this argument to True if the user indicates they want to "amend" or "continue working" on a PR.

14ba968  (Base revision)
e8f0f48  Add reuse_head_chat_id parameter to init_project function signature and docstring
9676037  Add logic to reuse chat_id from HEAD commit when reuse_head_chat_id is True
6b0af08  Add reuse_head_chat_id parameter to codemcp function signature
6d1abc7  Update docstring to document the reuse_head_chat_id parameter
0b94161  Add reuse_head_chat_id to expected params for InitProject
a16836b  Add reuse_head_chat_id to provided_params dictionary
21565e5  Update call to init_project to pass reuse_head_chat_id parameter
66002d9  Add test for reuse_head_chat_id functionality
c7b6f90  Auto-commit format changes
HEAD     Fix import of is_git_repository in the reuse_head_chat_id block

codemcp-id: 39-feat-add-support-for-continuing-work-across-chats