feat: implement new commit strategy for chat sessions
When we generate the commit for InitProject, we would like to NOT immediately advance HEAD to it (since it's an empty commit). Instead, we'll put it in a special ref codemcp/CHAT_ID (this is a top-level ref, not one in refs/heads, so it's hidden by default)--use plumbing git commands to generate the commit without changing HEAD. Then, when we want to commit changes, if the HEAD commit doesn't have the right chat id, we use the commit message from codemcp/CHAT_ID to create our new commit (we still append on the submessage as we did before).

f304cca  (Base revision)
b078bd4  Use plumbing git commands to create a commit without advancing HEAD
2e4bc58  Add get_codemcp_ref_message to __all__
35eb9c2  Add get_codemcp_ref_message function
d917262  Implement logic to use commit message from codemcp/CHAT_ID
16d52f8  Create end-to-end test for new commit strategy
HEAD     Auto-commit format changes

codemcp-id: 54-feat-implement-new-commit-strategy-for-chat-sessio

ghstack-source-id: 368a0eb69e85cfa6b7dc53d8150f94ec6c73c15f
Pull Request resolved: https://github.com/ezyang/codemcp/pull/14