refactor: change codemcp-id commit message handling to use regex
Let's redo how addition of codemcp-id / detection to commit message works. Right now, we have logic where we parse the commit message into a dict and directly read/edit the dict. In some other chats, we tried fixing a bug in this representation but ran into too many edge cases. What we are going to do instead is something even dumber: to read, we're going to regex for the last occurrence of "codemcp-id: XXX\n". To write, we're going to just paste "codemcp-id: XXX" at the end of the commit message with a double newline separating it from the regular body. This should work and be robust to what ghstack generates:

```
Subject

Foo desc
Bar bar

codemcp-id: 10-blah

Signed-off-by: foobar
ghstack-id: blahblahblah
```

Right now, codemcp-id is not properly detected, but with a backwards regex (getting the LAST occurrence) it would be.

321f63d  (Base revision)
89189a9  Update get_head_commit_chat_id to use regex pattern for finding the last occurrence of codemcp-id
20d5544  Update append_metadata_to_message to handle codemcp-id by directly appending it at the end
64755ee  Update get_ref_commit_chat_id to use regex pattern for finding the last occurrence of codemcp-id
0e2b896  Fix append_metadata_to_message to match expected test output format
a650858  Fix append_metadata_to_message to exactly match expected test output format
HEAD     Auto-commit format changes

codemcp-id: 61-refactor-change-codemcp-id-commit-message-handling

ghstack-source-id: 0e622b27aec5a52f58899fd7a93c16561285616f
Pull Request resolved: https://github.com/ezyang/codemcp/pull/18