refactor: change codemcp-id commit message handling to use regex
Let's redo how addition of codemcp-id / detection to commit message works. Right now, we have logic where we parse the commit message into a dict and directly read/edit the dict. In some other chats, we tried fixing a bug in this representation but ran into too many edge cases. What we are going to do instead is something even dumber: to read, we're going to regex for the last occurrence of "codemcp-id: XXX\n". To write, we're going to just paste "codemcp-id: XXX" at the end of the commit message with a double newline separating it from the regular body. This should work and be robust to what ghstack generates:

```
Subject

Foo desc
Bar bar

codemcp-id: 10-blah

Signed-off-by: foobar
ghstack-id: blahblahblah
```

Right now, codemcp-id is not properly detected, but with a backwards regex (getting the LAST occurrence) it would be.

d7ec0d1  (Base revision)
3b38425  Add test for complex commit message with codemcp-id in the middle
1a6e812  Add re module import to fix test
HEAD     Auto-commit format changes

codemcp-id: 61-refactor-change-codemcp-id-commit-message-handling

ghstack-source-id: f2cd654c6099c8d7a6b191f4dc85a54bc74e5e9b
Pull Request resolved: https://github.com/ezyang/codemcp/pull/19