refactor: use marker tokens for commit list in git_message.py
Let's refactor update_commit_message_with_description in git_message.py. Right now, it directly bashes the added text into the string. Instead, what we are going to do is define marker tokens "```git-revs" and "```" which demarcate where the list of commits occurs. We always put the list of commits in here. Now we are going to implement the "add a commit" algorithm like this: 1. parse out the contents between the marker tokens, 2. parse the internals (hash + space + description), 3. add ours (remember this involves filling in the hash for the last entry, and adding a new entry with HEAD), 4. reformat it back into the format, replacing it between the marker tokens. Importantly, the market tokens means that we have more robust detection of where the rev list occurs.

5f71037  (Base revision)
70ce9a7  Refactor update_commit_message_with_description to use marker tokens
571f0c1  Auto-commit format changes
14d7540  Fix compatibility with old commit message format
a8b3b34  Fix syntax errors in git_message.py
HEAD     Auto-commit format changes

codemcp-id: 70-refactor-use-marker-tokens-for-commit-list-in-git-