fix: add START/END_MARKER tokens to first commit
There's a bug in git_message.py regarding START_MARKER and END_MARKER tokens. Specifically, on the very first time we make a write (and insert our first git-revs list; previously there was none), we don't properly put it start/end markers. Modify an existing write file e2e test to test the FULL contents of the commit message of the first commit that is generated, fix the bug, and verify that the expected message has a ```git-rev marker in it.

codemcp-id: 78-fix-add-start-end-marker-tokens-to-first-commit

```git-revs
610f6c8  (Base revision)
e828709  Create test for InitProject and WriteFile interaction
7073534  Improve debug output for commit message inspection
785d4a0  Use direct git log command for debugging
2e34e41  Add test for git reference markers in commit messages
8c2b098  Fix create_commit_reference to use commit_hash=None to ensure markers
69b7e4c  Fix init_project to use update_commit_message_with_description
5997ef0  Test with null custom_message to let create_commit_reference generate it
1f2aadc  Fix commit message formatting in create_commit_reference
HEAD     Auto-commit format changes
```

ghstack-source-id: 484bb5d2f600a6ff64d4eee89db48d7b05c3ecfc
Pull Request resolved: https://github.com/ezyang/codemcp/pull/32