fix: maintain commit message structure when updating git-revs
We have a bug in git_message.py. Let's suppose we initially generate a commit like this (ignore leading indents in the examples below):

```
    feat: create foo.txt file

    Add a file foo.txt with contents foo

    ```git-revs
be899b1  (Base revision)
52663aa  Add test cases for update_commit_message_with_description function
fa0aa75  Stop stripping codemcp-id from the message
cb9cdcc  Only add codemcp-id if it doesn't already exist in the message
HEAD     Auto-commit lint changes
```

    codemcp-id: 79-feat-create-foo-txt-file
```

A third party tool may add some extra metadata at the bottom:

```
    feat: create foo.txt file

    Add a file foo.txt with contents foo

    ```git-revs
    be899b1  (Base revision)
    HEAD     Add foo.txt with content 'foo'
    ```

    codemcp-id: 79-feat-create-foo-txt-file

    ghstack-source-id: 1d28f5b40268b261d658603701a2e55ab545e7f5
    Pull Request resolved: https://github.com/ezyang/codemcp/pull/30
```

Now, the next time we amend the commit, our metadata information jumps back to the bottom:

```
    feat: create foo.txt file

    Add a file foo.txt with contents foo

    ghstack-source-id: 1d28f5b40268b261d658603701a2e55ab545e7f5
    Pull Request resolved: https://github.com/ezyang/codemcp/pull/30

    ```git-revs
    be899b1  (Base revision)
    3245873  Add foo.txt with content 'foo'
    HEAD     Add bar.txt with content 'bar'
    ```

    codemcp-id: 79-feat-create-foo-txt-file
```

This should not happen: (1) when there's already a codemcp-id, we should not rewrite it or delete it, we should leave it as is (this id never changes), and (2) when we update git-revs, we should leave the block where it is, and only change the inside content (so everything else stays relatively the same. So we expect:

```
    feat: create foo.txt file

    Add a file foo.txt with contents foo

    ```git-revs
    be899b1  (Base revision)
    3245873  Add foo.txt with content 'foo'
    HEAD     Add bar.txt with content 'bar'
    ```

    codemcp-id: 79-feat-create-foo-txt-file

    ghstack-source-id: 1d28f5b40268b261d658603701a2e55ab545e7f5
    Pull Request resolved: https://github.com/ezyang/codemcp/pull/30
```

Write a test for this behavior and then fix it.

codemcp-id: 80-fix-maintain-commit-message-structure-when-updatin
ghstack-source-id: 1d3b6198159e8a580e7063c4726c78cea7c49e2e