refactor: Simplify WriteFile commit message generation
test_write_file in test_write_file.py is currently failing. The test is correct but the implementation is wrong. The problem should be related to how WriteFile generates commit messages in the git helper modules. I think the code in that area is a bit messy. Rather than directly try to fix the test, can you examine the relevant functions and see if there's a refactor to make it simpler, that would also fix the test?

```git-revs
836dc45  (Base revision)
5561d98  Auto-commit accept changes
4f576b0  Auto-commit format changes
HEAD     Auto-commit lint changes
```

codemcp-id: 117-refactor-simplify-writefile-commit-message-generat

ghstack-source-id: 4460dabef1a008ebfd924636ff9568c3ba426d67
Pull Request resolved: https://github.com/ezyang/codemcp/pull/60