refactor: Simplify WriteFile commit message generation
test_write_file in test_write_file.py is currently failing. The test is correct but the implementation is wrong. The problem should be related to how WriteFile generates commit messages in the git helper modules. I think the code in that area is a bit messy. Rather than directly try to fix the test, can you examine the relevant functions and see if there's a refactor to make it simpler, that would also fix the test?

```git-revs
e778097  (Base revision)
0802104  Simplify commit message update logic
HEAD     Simplify commit message generation in git_commit.py
```

codemcp-id: 117-refactor-simplify-writefile-commit-message-generat

ghstack-source-id: 241b14034e6495b02b250d866e5c9b9101520129
Pull Request resolved: https://github.com/ezyang/codemcp/pull/58