fix: investigate Windows line ending issue in WriteFile
Users on Windows reported that when codemcp writes files it adds extra spaces in between lines of code. I'm not too sure why this happens. Read the code for WriteFile tool and carefully think about might occur. You cannot test the problem here as we are not on Windows, so be confident about your fix or discuss it with me.

```git-revs
000cccb  (Base revision)
846792c  Fix Windows line ending issue by using binary mode for file write
HEAD     Auto-commit lint changes
```

codemcp-id: 163-fix-investigate-windows-line-ending-issue-in-write
ghstack-source-id: 0effedb63e7704a2be83e8bd9438aa8d90e0c5ed
Pull-Request-resolved: https://github.com/ezyang/codemcp/pull/148