refactor: remove codemcp main entry point and update tests
Delete codemcp.main.codemcp, and then fix tests to deal with it. In general, if code was calling codemcp, it should now be calling the internal tool it wanted to use directly. If the tool was not known, there's a dispatch function *dispatch*to_subtool available in the testing class but this should never be used in the actual production code.

```git-revs
5301e40  (Base revision)
7dfc6f4  Remove the codemcp function from main.py
0084795  Update comment in _dispatch_to_subtool method
1ef03c4  Remove comment about not using codemcp.main.codemcp
HEAD     Remove codemcp import from __init__.py
```

codemcp-id: 298-refactor-remove-codemcp-main-entry-point-and-updat